refactor: /counsellor/modules route
